requirements.txt
annotated-types==0.7.0
azure-common==1.1.28
azure-core==1.35.0
azure-functions==1.23.0
azure-identity==1.24.0
azure-mgmt-core==1.6.0
azure-mgmt-resource==24.0.0
azure-mgmt-subscription==3.1.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
idna==3.10
isodate==0.7.2
MarkupSafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.3.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
